requirements.txt
fastapi[standard]>=0.111.0
uvicorn[standard]>=0.29.0
sqlmodel>=0.0.19
revpimodio2
numpy
pyserial
